Revert "✨ Refactor architecture and add DI"
