+++ sld-api-backend/core/provider.py
@@ -72,7 +72,6 @@ class ProviderActions:
     """
     This class contains the typical methods of a deployment
     """
-
     def plan(
         name: str,
         stack_name: str,
