+++ sld-api-backend/src/activityLogs/api/v1/activity_logs.py
@@ -1,9 +1,9 @@
+from src.activityLogs.infrastructure import repositories  as crud_activity
+from src.users.infrastructure import repositories as crud_users
+from src.users.domain.entities import users as schemas_users
 from fastapi import APIRouter, Depends, HTTPException
 from security import deps
 from sqlalchemy.orm import Session
-from src.activityLogs.infrastructure import repositories as crud_activity
-from src.users.domain.entities import users as schemas_users
-from src.users.infrastructure import repositories as crud_users
 
 router = APIRouter()
