+++ sld-api-backend/src/activityLogs/domain/entities/activity.py
