+++ sld-api-backend/src/aws/api/v1/aws.py
@@ -1,11 +1,11 @@
+from src.activityLogs.infrastructure import repositories  as crud_activity
+from src.aws.infrastructure import repositories as crud_aws
+from src.aws.domain.entities import aws as schemas_aws
+from src.users.infrastructure import repositories as crud_users
+from src.users.domain.entities import users as schemas_users
 from fastapi import APIRouter, Depends, HTTPException, Response
 from security import deps
 from sqlalchemy.orm import Session
-from src.activityLogs.infrastructure import repositories as crud_activity
-from src.aws.domain.entities import aws as schemas_aws
-from src.aws.infrastructure import repositories as crud_aws
-from src.users.domain.entities import users as schemas_users
-from src.users.infrastructure import repositories as crud_users
 
 router = APIRouter()
