+++ sld-api-backend/src/aws/domain/entities/aws.py
