+++ sld-api-backend/src/aws/infrastructure/models.py
@@ -4,7 +4,6 @@ from config.database import Base
 from sqlalchemy import (Column, DateTime, Integer, String, Text,
                         UniqueConstraint)
 
-
 class Aws_provider(Base):
     __tablename__ = "aws_provider"
     id = Column(Integer, primary_key=True, index=True)
