+++ sld-api-backend/src/aws/infrastructure/repositories.py
@@ -1,9 +1,9 @@
 import datetime
 
 import src.aws.infrastructure.models as models
+from src.aws.domain.entities import aws as schemas_aws
 from security.vault import vault_decrypt, vault_encrypt
 from sqlalchemy.orm import Session
-from src.aws.domain.entities import aws as schemas_aws
 
 
 @vault_encrypt
