+++ sld-api-backend/src/azure/domain/entities/azure.py
@@ -1,6 +1,8 @@
 from pydantic import BaseModel, constr
 
 
+
+
 class AzureBase(BaseModel):
     squad: constr(strip_whitespace=True)
     environment: constr(strip_whitespace=True)
