+++ sld-api-backend/src/azure/infrastructure/repositories.py
@@ -2,10 +2,10 @@ import datetime
 import imp
 
 import src.azure.infrastructure.models as models
+from src.azure.domain.entities import azure as schemas_azure
 from security.vault import vault_decrypt, vault_encrypt
 from sqlalchemy import exc
 from sqlalchemy.orm import Session
-from src.azure.domain.entities import azure as schemas_azure
 
 
 @vault_encrypt
