+++ sld-api-backend/src/deploy/api/v1/schedule.py
@@ -1,3 +1,6 @@
+from src.tasks.infrastructure import repositories as crud_tasks
+from src.users.infrastructure import repositories as crud_users
+from src.users.domain.entities import users as schemas_users
 from fastapi import (APIRouter, BackgroundTasks, Depends, HTTPException,
                      Response, status)
 from helpers.get_data import (check_cron_schedule, check_squad_user, deploy,
@@ -5,13 +8,10 @@ from helpers.get_data import (check_cron_schedule, check_squad_user, deploy,
 from helpers.push_task import (async_schedule_add, async_schedule_delete,
                                async_schedule_get, async_schedule_list,
                                async_schedule_update)
-from security import deps
-from sqlalchemy.orm import Session
 from src.deploy.domain.entities import schedule as schemas_schedule
 from src.deploy.infrastructure import repositories as crud_deploys
-from src.tasks.infrastructure import repositories as crud_tasks
-from src.users.domain.entities import users as schemas_users
-from src.users.infrastructure import repositories as crud_users
+from security import deps
+from sqlalchemy.orm import Session
 
 router = APIRouter()
