+++ sld-api-backend/src/deploy/domain/entities/deploy.py
@@ -3,6 +3,7 @@ from typing import List, Optional
 from pydantic import BaseModel, Field, constr
 
 
+
 class DeployBase(BaseModel):
     name: constr(strip_whitespace=True)
     stack_name: constr(strip_whitespace=True)
