+++ sld-api-backend/src/gcp/domain/entities/gcp.py
