+++ sld-api-backend/src/tasks/api/v1/tasks.py
@@ -2,20 +2,19 @@ import ast
 
 from celery.result import AsyncResult
 from config.celery_config import celery_app
+from src.tasks.infrastructure import repositories as crud_tasks
+from src.users.infrastructure import repositories as crud_users
+from src.users.domain.entities import users as schemas_users
 from fastapi import APIRouter, Depends, HTTPException
 from security import deps
 from sqlalchemy.orm import Session
-from src.tasks.infrastructure import repositories as crud_tasks
-from src.users.domain.entities import users as schemas_users
-from src.users.infrastructure import repositories as crud_users
 
 router = APIRouter()
 
 
 @router.delete("/id/{task_id}")
 async def get_task_by_id(
-    task_id: str,
-    current_user: schemas_users.User = Depends(deps.get_current_active_user),
+    task_id: str, current_user: schemas_users.User = Depends(deps.get_current_active_user)
 ):
     result = celery_app.control.revoke(task_id, terminate=True)
     return {"result": f"REVOKE {task_id}"}
@@ -23,8 +22,7 @@ async def get_task_by_id(
 
 @router.get("/id/{task_id}")
 async def get_task_by_id(
-    task_id: str,
-    current_user: schemas_users.User = Depends(deps.get_current_active_user),
+    task_id: str, current_user: schemas_users.User = Depends(deps.get_current_active_user)
 ):
     a = AsyncResult(task_id)
     cstr = str(a.info)
