+++ sld-api-backend/src/tasks/infrastructure/models.py
@@ -1,7 +1,8 @@
 import datetime
 
 from config.database import Base
-from sqlalchemy import Column, DateTime, Integer, String
+from sqlalchemy import (Column, DateTime, Integer, String)
+   
 
 
 class Tasks(Base):
