+++ sld-api-backend/src/users/api/v1/auth.py
@@ -2,11 +2,10 @@ from typing import Any
 
 from fastapi import APIRouter, Depends
 from fastapi.security import OAuth2PasswordRequestForm
-from security import deps
-from src.users.application.tokens import UserExist
-from sqlalchemy.orm import Session
 from src.users.domain.entities import users as schemas_users
-from src.users.infrastructure import repositories as crud_users
+from security import deps
+from security.tokens import validate_user
+from sqlalchemy.orm import Session
 
 router = APIRouter()
 
@@ -18,9 +17,7 @@ def login_access_token(
     """
     OAuth2 compatible token login, get an access token for future requests
     """
-    user_db = crud_users.get_user_by_username(db, username=user.username)
-    validate = UserExist(user_db, user.username, user.password)
-    return validate.validate_user()
+    return validate_user(db, user.username, user.password)
 
 
 @router.post("/access-token-json", response_model=schemas_users.Token)
@@ -30,6 +27,4 @@ def login_access_token_json(
     """
     OAuth2 compatible token login, get an access token for future requests
     """
-    user_db = crud_users.get_user_by_username(db, username=user.username)
-    validate = UserExist(user_db, user.username, user.password)
-    return validate.validate_user()
+    return validate_user(db, user.username, user.password)
