+++ sld-api-backend/src/users/domain/entities/users.py
@@ -59,6 +59,7 @@ class TokenData(BaseModel):
     username: str = None
 
 
+
 class PlanCreate(BaseModel):
     name: constr(strip_whitespace=True)
     stack_name: constr(strip_whitespace=True)
