+++ sld-api-backend/src/variables/api/v1/variables.py
@@ -1,11 +1,11 @@
+from src.deploy.infrastructure import repositories as crud_deploys
+from src.stacks.infrastructure import repositories as crud_stacks
+from src.users.infrastructure import repositories as crud_users
+from src.users.domain.entities import users as schemas_users
 from fastapi import APIRouter, Depends, HTTPException
 from helpers.get_data import check_squad_user
 from security import deps
 from sqlalchemy.orm import Session
-from src.deploy.infrastructure import repositories as crud_deploys
-from src.stacks.infrastructure import repositories as crud_stacks
-from src.users.domain.entities import users as schemas_users
-from src.users.infrastructure import repositories as crud_users
 
 router = APIRouter()
